+++ android/src/main/java/vn/hunghd/flutterdownloader/FlutterDownloaderPlugin.java
@@ -44,7 +44,7 @@ public class FlutterDownloaderPlugin implements MethodCallHandler, FlutterPlugin
     private TaskDao taskDao;
     private Context context;
     private long callbackHandle;
-    private int stepUpdate;
+    private int step;
     private int debugMode;
     private int ignoreSsl;
 
@@ -127,7 +127,7 @@ public class FlutterDownloaderPlugin implements MethodCallHandler, FlutterPlugin
                         .putBoolean(DownloadWorker.ARG_OPEN_FILE_FROM_NOTIFICATION, openFileFromNotification)
                         .putBoolean(DownloadWorker.ARG_IS_RESUME, isResume)
                         .putLong(DownloadWorker.ARG_CALLBACK_HANDLE, callbackHandle)
-                        .putInt(DownloadWorker.ARG_STEP_UPDATE, stepUpdate)
+                        .putInt(DownloadWorker.ARG_STEP, step)
                         .putBoolean(DownloadWorker.ARG_DEBUG, debugMode == 1)
                         .putBoolean(DownloadWorker.ARG_IGNORESSL, ignoreSsl == 1)
                         .putBoolean(DownloadWorker.ARG_SAVE_IN_PUBLIC_STORAGE, saveInPublicStorage)
@@ -160,7 +160,7 @@ public class FlutterDownloaderPlugin implements MethodCallHandler, FlutterPlugin
     private void registerCallback(MethodCall call, MethodChannel.Result result) {
         List args = (List) call.arguments;
         callbackHandle = Long.parseLong(args.get(0).toString());
-        stepUpdate = Integer.parseInt(args.get(1).toString());
+        step = Integer.parseInt(args.get(1).toString());
         result.success(null);
     }